Moved ApplicationChoiceListComponent
