+++ app/components/candidate_interface/application_choice_list_component.html.erb
