+++ app/views/candidate_interface/continuous_applications_choices/index.html.erb
@@ -63,7 +63,7 @@
       <% end %>
     <% end %>
 
-    <%= render CandidateInterface::ContinuousApplications::ApplicationChoiceListComponent.new(application_form: current_application, application_choices: @application_choices, current_tab_name: params[:current_tab_name]) %>
+    <%= render CandidateInterface::ApplicationChoiceListComponent.new(application_form: current_application, application_choices: @application_choices, current_tab_name: params[:current_tab_name]) %>
   <% end %>
   </div>
 </div>
